requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==4.2.13
django-ipware==7.0.1
djangorestframework==3.15.2
future==1.0.0
geocoder==1.38.1
idna==3.7
python-ipware==3.0.0
ratelim==0.1.6
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
